+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * Add support for Content-Security-Policy Level 2 directives
   * Add support for Content-Security-Policy Level 2 signatures (nonce and message digest)
   * Deprecate report-uri as an array. Property is now a scalar
+  * Deprecate encrypted cookie support du to high coupling to mcrypt deprecated extension
   * Drop backward-compatibility with first deprecated CSP configuration
 
 ### 1.10.0 (2016-02-23)